requirements.txt
# FastAPI и его сервер
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Работа с PostgreSQL
psycopg2-binary==2.9.9

# Работа с .env файлами и настройками
pydantic-settings==2.2.1
python-dotenv==1.0.1